requirements.txt
streamlit==0.88.0
numpy==1.19.5   
pandas==1.3.2   
pillow==8.3.1
matplotlib==3.4.3
tensorflow ==2.6.0    
opencv-python==4.5.3.56